requirements.txt
pandas
streamlit==1.37.0
plotly
scikit-learn
altair==5.3.0
numpy
supabase
python-dotenv
prophet
xgboost
statsmodels
cmdstanpy
openpyxl
sendgrid
fastapi
uvicorn[standard]
python-multipart
httpx
requests
